More CRUD progress on furnituring

+++ BookingSystemBackEnd/BookingSystem/Models/BLL/Furnituring.cs
@@ -9,7 +9,7 @@ namespace BookingSystem.Models
     public class Furnituring
     {
         [Required(ErrorMessage = "FurnituringId is required.")]
-        [Range(0, int.MaxValue, ErrorMessage = "FurnituringId is out of range.")]
+        [Range(0, Int16.MaxValue, ErrorMessage = "FurnituringId is out of range.")]
         public int FurnituringId { get; set; }
 
         [StringLength(50, ErrorMessage = "Name must not exceed 50 chars.")]
